+++ src/test/java/dk/kb/discover/SolrServiceTest.java
@@ -1,7 +1,13 @@
 package dk.kb.discover;
 
 import dk.kb.discover.api.v1.impl.DsDiscoverApiServiceImpl;
+import dk.kb.discover.config.ServiceConfig;
+import dk.kb.discover.util.SolrParamMerger;
+import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
+
+import java.io.IOException;
+import java.net.URI;
 
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -119,4 +125,52 @@ class SolrServiceTest {
         assertThrows(IllegalArgumentException.class,
                 () -> SolrService.removePrefixedFilters(response, prefix,"xml"));
     }
+
+    @Test
+    @Tag("integration")
+    void suggestTestNotAvailable() throws IOException {
+        // Integration test towards devel env. Remember to update aegis before running this.
+        ServiceConfig.initialize("src/test/resources/ds-discover-integration-test.yaml");
+        String suggestDictionary = "radiotv_title_suggest";
+        // no suggestions should be available for this query.
+        String suggestQuery = "Palle Lauring";
+        int suggestCount = 5;
+        String wt = "json";
+        SolrService solr = SolrManager.getSolrService("ds");
+
+        String filteredResponse = solr.suggest(suggestDictionary, suggestQuery, suggestCount, wt);
+
+        assertTrue(filteredResponse.contains("\"suggest\" : {\n" +
+                "    \"radiotv_title_suggest\" : {\n" +
+                "      \"Palle Lauring\" : {\n" +
+                "        \"numFound\" : 0,\n" +
+                "        \"suggestions\" : [ ]\n" +
+                "      }\n" +
+                "    }\n" +
+                "  }"));
+    }
+
+    @Test
+    @Tag("integration")
+    void suggestTestMoreThanFiveAvailable() throws IOException {
+        // Integration test towards devel env. Remember to update aegis before running this.
+        ServiceConfig.initialize("src/test/resources/ds-discover-integration-test.yaml");
+        String suggestDictionary = "radiotv_title_suggest";
+        // no suggestions should be available for this query.
+        String suggestQuery = "tes";
+        int suggestCount = 5;
+        String wt = "json";
+        SolrService solr = SolrManager.getSolrService("ds");
+
+        String filteredResponse = solr.suggest(suggestDictionary, suggestQuery, suggestCount, wt);
+
+        System.out.println(filteredResponse);
+
+        assertTrue(filteredResponse.contains("\"suggest\" : {\n" +
+                "    \"radiotv_title_suggest\" : {\n" +
+                "      \"tes\" : {\n" +
+                "        \"numFound\" : 5,\n"));
+    }
+
+
 }