+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Changed
+
+### Added 
+- Added filtering to suggest component, to make it only suggest documents, that the caller are able to query for.
 - make all loggers static
 
 ## [1.4.1](https://github.com/kb-dk/ds-discover/releases/tag/ds-discover-1.4.1) - 2024-10-15